passing props to HOC and deleted files append with LifeCycle
