+++ src/components/Cockpit/Cockpit.js
@@ -1,4 +1,5 @@
-import React from "react";
+import React, { useEffect } from "react";
+
 import styled from "styled-components";
 import classes from "./Cockpit.module.css";
 
@@ -14,23 +15,30 @@ const StyledButton = styled.button`
   }
 `;
 
-const cockpit = (props) => {
+const Cockpit = props => {
+
+ useEffect(() => {
+    console.log("[Cockpit.js] useEffect"); 
+     
+  } ); 
+  
+  
   let btnClass = classes.simple;
-  if (props.personlength>=2) {
+  if (props.personlength<=2) {
     btnClass = classes.Violet;
   }
- 
   return (
     <div className={classes.Cockpit}>
-    {/* Adding cockpit styling inside a div is necassary to apply only cockpit css scope in below items else App scope will be overridden on it */}
+      {/* Adding cockpit styling inside a div is necassary to apply only cockpit css scope in below items else App scope will be overridden on it */}
       <h1>{props.title}</h1>
       <p>This is really working</p>
       <button className={btnClass} onClick={() => props.clicked("from button")}>
-        switch name 
+        switch name
       </button>
-      <StyledButton onClick={props.toogled}> toggle </StyledButton>
+      <StyledButton onClick={props.toogled}> togglePerson </StyledButton>
     </div>
   );
 };
 
-export default cockpit;
+export default React.memo(Cockpit);
+//export default Cockpit;
