+++ src/components/Persons/Person/Person.js
@@ -1,31 +1,26 @@
-import React from "react";
-import "./Person.css";
-import styled from "styled-components";
+import React, { Component } from "react";
+import classes from "./Person.module.css";
+import Aux from "../../../hoc/Auxilary.js";
+import WithClass from "../../../hoc/WithClass.js";
 
-const StyleDiv = styled.div`
-  width: 60%;
-  margin: auto;
-  border: 3px solid #eeeeee;
-  box-shadow: 0 2px 3px #cccccc;
-  padding: 16px;
-  text-align: center;
-  @media (min-width: 500px) {
-    width: 450px;
+
+class Person  extends Component {
+  render() {
+    //console.log("[Person_LC.js] rendering...");
+    return (
+      <Aux>
+        <p onClick={this.props.click}>
+          I'm {this.props.name} and I am {this.props.age} years old!
+        </p>
+        <p>{this.props.children}</p>
+        <input
+          type="text"
+          onChange={this.props.changed}
+          value={this.props.name}
+        />
+      </Aux>
+    );
   }
-`;
- 
-const person = props => {
-  console.log("[Person.js] rendering");
-  return (
-    //<div className="Person" style={style}>
-    <StyleDiv>
-      <p onClick={props.click}>
-        I am {props.name} having age {props.age} person{" "}
-      </p>
-      <p>{props.children}</p>
-      <input type="text" onChange={props.changed} value={props.name}></input>
-    </StyleDiv>
-  );
-};
+}
 
-export default person;
+export default WithClass(Person,classes.Person);
