--- src/components/Persons/Person/PersonLifeCycle.js
@@ -1,25 +1,0 @@
-import React, { Component } from "react";
-import classes from "./Person.module.css";
-import Aux from "../../../hoc/Auxilary.js";
-
-
-class PersonLifeCycle extends Component {
-  render() {
-    //console.log("[Person_LC.js] rendering...");
-    return (
-      <div className={classes.Person}>
-        <p onClick={this.props.click}>
-          I'm {this.props.name} and I am {this.props.age} years old!
-        </p>
-        <p>{this.props.children}</p>
-        <input
-          type="text"
-          onChange={this.props.changed}
-          value={this.props.name}
-        />
-      </div>
-    );
-  }
-}
-
-export default PersonLifeCycle;
