+++ src/components/Persons/Persons.js
@@ -1,18 +1,23 @@
-import React from "react";
+import React, { PureComponent } from "react";
 import Person from "./Person/Person.js";
 
-const Persons = (prop) =>{
-console.log('[Persons.js] rendering');
-return prop.persons.map((person, index) => {
-  return (
-    <Person
-      click={() => prop.clicked(index)}
-      name={person.name}
-      age={person.age}
-      key={person.id} // Add key to improve performance for list. Since while rendering react compares new virtual DOM with the old one and for this it needs an unique identifier which states what got changed
-      changed={event => prop.changed(event, person.id)}
-    />
-  );
-})}
-
+class Persons  extends PureComponent {
+  render() {
+    //console.log('[Persons_LC.js] rendering...');
+    return this.props.persons.map((person, index) => {
+      return (
+        <Person
+          click={() => this.props.clicked(index)}
+          name={person.name}
+          age={person.age}
+          key={person.id}
+          changed={event => this.props.changed(event, person.id)}
+        />
+      );
+    });
+  }
+}
 export default Persons;
+ 
+
+
