+++ src/hoc/WithClass.js
@@ -1,9 +1,10 @@
 import React from "react";
+ 
 //Its a normal function which returns a functional component
 const withClass = (WrappedComponent, className) => {
   return props => (
     <div className={className}>
-      <WrappedComponent />
+      <WrappedComponent {...props}/>
     </div>
   );
 };
